pyproject.toml
[project]
name = "nanobrowser"
version = "0.1.0a1"
description = "A multi-agent browser automation tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "langchain>=0.3.12",
    "langchain-anthropic>=0.3.1",
    "langchain-openai>=0.2.14",
    "pdfplumber>=0.11.4",
    "playwright>=1.49.1",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "websockets>=14.1",
]

[project.scripts]
nanobrowser = "nanobrowser.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nanobrowser"]

[tool.hatch.metadata]
allow-direct-references = true

